feat(sink/webhook): use utils retry_operation

+++ src/sinks/webhook/setup.rs
@@ -21,7 +21,7 @@ pub struct Config {
     pub headers: Option<HashMap<String, String>>,
     pub timeout: Option<u64>,
     pub error_policy: Option<ErrorPolicy>,
-    pub max_retries: Option<usize>,
+    pub max_retries: Option<u32>,
     pub backoff_delay: Option<u64>,
 }
 
@@ -49,7 +49,7 @@ fn build_headers_map(config: &Config) -> Result<HeaderMap, Error> {
     Ok(headers)
 }
 
-const DEFAULT_MAX_RETRIES: usize = 20;
+const DEFAULT_MAX_RETRIES: u32 = 20;
 const DEFAULT_BACKOFF_DELAY: u64 = 5_000;
 
 impl SinkProvider for WithUtils<Config> {
